Create wildbill.txt
Profile link: www.BonusBots.com/wildbill.txt

» Doodle tweak (run Doodle .ppl cash game profile at the same time, even for tournaments)

» Buyin for 50 BB’s in cash games! (No more, no less – do not play shallow or deep stack tables, stick to the normal 100 BB max buyin games, especially 6-max tables).

» Do not take profits or use a stop-loss! Do not hop tables! Just let it keep playing for 6-10 hours.

This is my latest profile, which is really fun to watch. I modeled it after a friend of mine who is also fun to watch play; in fact he absolutely makes the game at any table he sits down at. As you may expect, it pushes a lot. I love it. But if the idea of preflop shoving with 87 suited or 55 makes you shudder, this profile probably is not for you. My friend who plays wild usually has a large stack at some point in the session, being as he gets a lot of action. Similarly, it is a rare session that this profile does not triple up its starting stack at some point (albeit sometimes that doesn’t get it even on the session yet). 

I started developing this as a Cake tournament profile, wanting to counteract the issue of the bot not having the betsize or potsize information available when an opponent pushes. But it has mostly been tested in 6-max cash games and 1-table SNGs there. It has done well, with the current version providing a positive overall return in my testing across all game types (including DONs) except for MTTs, which I have not had good results at yet. It seems to do well in cash games as high as .05/.10 NL however and makes the money quite often in regular 1-table SNGs. It has not been tested in full ring cash games so I would stick to 6-max tables.

While developing this profile I discovered that Cake is not such a bad place to go botting! I would be curious to see how it does at the micro-limit 6-max NL tables on iPoker. I am willing to keep developing this puppy, so feel free to post hand histories here – however please do not suggest less pushing because this is, after all, Wild Bill. Here are some of the profile features:

• Tightens up with a stack of 150+. Your opponents don’t realize this! So if you start with 50 BB’s as recommended, after you triple up it then only pushes with quality - but continues to get loose calls from your opponents who have seen all the earlier wild play. I have seen Wild Bill reach stack sizes of 450+ quite often in cash games, still getting ridiculously loose calls for entire stacks from opponents who are acting on the wider small-stack pushing range.

• Tighter tournament play. Many code lines include a bigblindsize > 10 stipulation which makes it less wild in tournaments.

• No preflop min-raising. All preflop open-raises are either by 2 BB’s (making it 3 total) or flat-out pushes.

• Flop C-betting is 70% x 75%. That is, 70% of the pot and done 75% of the time randomly. However this is the remaining 75% of the time that it does not push on the flop, which it does with most good flopped hands and draws. With very strong flopped hands it will also sometimes bet 75% of the pot as a trap.

• Self-contained Preflop and Flop code. The Turn and River betting rounds are Doodle Tweaks and need the Doodle cash profile in the background (no need for tournament codes as this profile has them included).

• Usersteal and Userdefend user-variables for combat situations. By identifying our own preflop late-position stealing situations, and defense situations against late-position stealers, we then play better combat post-flop. By incorporating these user-defined variables, Wild Bill will do things like:
  -Float smallish c-bets on the flop (with any hand) from preflop late-position stealers then bet the Turn or River when they check
  -Reraise all-in on the flop randomly 30% of the time (with any hand) when we ourselves are the preflop late-position stealer and our flop C-bet gets raised by a not-huge amount (except when stacksize > 150 or bigblindsize > 10)
  -Occasionally push on the flop (with any hand) from preflop late-position stealers when they make a flop C-bet on non-scary boards (except when stacksize > 150 or bigblindsize > 10)

OK all you Wild Bills let’s get you seated and get some feedback on this thread!