Create green
This is a Doodle tweak for cash games that was inspired by one of the strategies described in Phil Gordon's Little Green Book. You will notice that it is very short so I am including the entire code in the post here. These are the main points:

• Pushes Liberally on the Flop ...with any decent draw (straight draws on non-flush boards or any 2+2 flush draw) plus any good hand (AK top pair on non-scary boards, QQ+ overpairs on non-scary boards, 2-pair hands on non-scary boards, and sets).

• Calls Liberally Preflop ...with drawing-type hands like small pocket pairs, suited aces, suited big cards, and connectors (suited or not) so that it gets more chances to push on the flop.

• That's it! This profile is not for the faint of heart. It will have big swings! Sometimes they are up however, and I have had days where I cashed in more than 10 times the max buyin after just a few hours of play using it. 

So please do not post bot-log examples questioning the play when you see it push with a straight draw! That is the whole intent of this profile! 

And yes you should always buyin for the table max! This would be a terrible short-stacking strategy IMO! If you do not buyin for the max expect to get killed with this.

The basic intent is to drive your opponents nuts. They will notice the liberal flop-pushing and resolve to call you down with any top pair hand. They will pick the wrong times to do that sometimes, and even when they pick the right time to do it, you will draw out on them sometimes and when that happens it can really put them on tilt.

I have had some good success with it at 6-max 25NL games on Bodog. However, don't ask me if this is a 25NL profile or a 10NL profile because stakes are irrelevant. It is game conditions that matter. Drives me nuts when I see people ask if a profile is for 10NL or 5NL as if stakes had anything to do with any of the codeline executions. 

Also, a good cash game profile should be able to play equally well at 10-max, 9-max, or 6-max tables as long as both full table tight play and short-handed aggressive play is coded for. I don't make profiles that are meant to sit out when tables get short, because it is better to keep on playing and hope it fills back up IMO, which is why the Doodle has a bunch of heads-up code included. That being said, in 6-max games you get more cards dealt and therefore have more chances to push and are likely to stand out from the crowd more for your liberal pushing, so I would expect it to do better in them, but who knows.