Create 6pack
This is another Doodle tweak, but obviously you can tweak any cash game profile with it as well. However, please note that the Doodle user-variables have been blended in to set true in this profile (as of version 10). If using a different cash game profile as a back-up which has it's own user-variables, you should manually blend those in. This might require emailing your profile creator to get them.

Style of Play:

I tried to make a LAG profile for 6-max cash tables that generally plays them the way I do personally. Lots of loose open-raising, and lots of random raising of one opponent on the flop, including some check-raising from first-position as the pre-flop aggressor. Also some random pushing on the flop with good hands + certain draws (only after an opponent bets or raises, so not nearly as much of this as the Little Green profile does).

I'd like to further develop this. Anyone who wants to help with that, please post bot-logs of hands played with your suggestions for improvement.